Crate boa_icu_provider

Boa’s boa_icu_provider exports the default data provider used by its Intl implementation.

Crate Overview

This crate exports the function buffer, which contains an extensive dataset of locale data to enable Intl functionality in the engine. The set of locales included is precisely the “modern” subset of locales in the Unicode Common Locale Data Repository.

If you need to support the full set of locales, you can check out the ICU4X guide about generating custom data providers. Boa supports plugging both BufferProviders or AnyProviders generated by the tool.

About Boa

Boa is an open-source, experimental ECMAScript Engine written in Rust for lexing, parsing and executing ECMAScript/JavaScript. Currently, Boa supports some of the language. More information can be viewed at Boa’s website.

Try out the most recent release with Boa’s live demo playground.

Boa Crates

boa_ast - Boa’s ECMAScript Abstract Syntax Tree.

boa_engine - Boa’s implementation of ECMAScript builtin objects and execution.

boa_gc - Boa’s garbage collector.

boa_interner - Boa’s string interner.

boa_parser - Boa’s lexer and parser.

boa_profiler - Boa’s code profiler.

boa_icu_provider - Boa’s ICU4X data provider.

boa_runtime - Boa’s WebAPI features.

Modules

baked 🔒

Functions

buffer

Gets the default data provider stored as a BufferProvider.

minimal

Gets a minimal data provider that is used when the intl feature of boa_engine is disabled.