Macro boa_icu_provider::baked::impl_any_provider

macro_rules! impl_any_provider {
    ($ provider : path) => { ... };
}

Implement [AnyProvider] on the given struct using the data hardcoded in this module. This allows the struct to be used with icu’s _any constructors.

This macro can only be called from its definition-site, i.e. right after include!-ing the generated module.

struct MyAnyProvider;
include!("/path/to/generated/mod.rs");
impl_any_provider(MyAnyProvider);