Crate boa_icu_provider

Boa’s boa_icu_provider exports the default data provider used by its Intl implementation.

Crate Overview

This crate exports the function buffer, which contains an extensive dataset of locale data to enable Intl functionality in the engine. The set of locales included is precisely the “modern” subset of locales in the Unicode Common Locale Data Repository.

If you need to support the full set of locales, you can check out the ICU4X guide about generating custom data providers. Boa supports plugging both BufferProviders or AnyProviders generated by the tool.

Statics

PROVIDER 🔒

Functions

buffer

Gets a data provider that is stored as a [BufferProvider]